requirements.txt
toml
psutil
pynvml
py-cpuinfo